Create README.md
Contactless Gesture Recognition using Air Canvas is a major project designed and implemented by our team 
which focuses on the primary feature of airwriting. The add-on features include Digital White Canvas, Virtual 
ink with a customizable colour palette, Optical character recognition (OCR), Duplication (Copy and Paste) and 
Screen capture. We have used Python as the primary programming language and Visual Code Editor as the 
platform for the implementation.